+++ front/src/states/index.tsx
@@ -1,1 +1,2 @@
-export { MessageRecoil } from './atoms';
+export type { Message } from './atoms';
+export { MessageRecoil, MessageListRecoil } from './atoms';
